Change the return type of wxKeyEvent::GetUnicodeKey from wxChar to wxUniChar
wxKeyEvent::GetUnicodeKey should return a single Unicode code point,
even if wxChar is 16 bit.

+++ src/gtk/window.cpp
@@ -1268,7 +1268,7 @@ bool wxWindowGTK::GTKDoInsertTextFromIM(const char* str)
 #if wxUSE_UNICODE
         event.m_uniChar = *pstr;
         // Backward compatible for ISO-8859-1
-        event.m_keyCode = *pstr < 256 ? event.m_uniChar : 0;
+        event.m_keyCode = *pstr < 256 ? (int)event.m_uniChar : 0;
         wxLogTrace(TRACE_KEYS, wxT("IM sent character '%c'"), event.m_uniChar);
 #else
         event.m_keyCode = (char)*pstr;
